Extract structs into their separate file
